requirements.txt
python-dotenv
python-telegram-bot
pillow
emoji
caribou